refactor blogs and concat user id

+++ FullStack-Part4-master/controllers/blogs.js
@@ -1,11 +1,11 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
+
 const Blog = require('../models/blog')
-// const User = require('../models/user')
-// const jwt = require('jsonwebtoken')
+const User = require('../models/user')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  // .populate('user', { username: 1, name: 1 })
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.status(200).send(blogs)
 })
 
@@ -15,27 +15,23 @@ blogRouter.post('/', async (req, res, next) => {
   if(data.title === undefined && data.url === undefined) {
     return res.status(400).end()
   } else {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    if (!req.token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
     const blog = new Blog ({
       title: data.title,
       author: data.author,
       url: data.url,
-      likes: data.likes
+      likes: data.likes,
+      user: user._id
     })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     res.status(201).json(savedBlog)
   }
-  // const decodedToken = jwt.verify(req.token, process.env.SECRET)
-  // if (!req.token || !decodedToken.id) {
-  //   return res.status(401).json({ error: 'token missing or invalid' })
-  // }
-  // const user = await User.findById()
-  /*blog.save()
-      .then(savedBlog => {
-        response.json(savedBlog.toJSON())
-      })
-      .catch(error => next(error))*/
-  /*const savedBlog = await blog.save()
-    response.json(savedBlog.toJSON())*/
 })
 
 blogRouter.get('/:id', async (req, res) => {
@@ -48,29 +44,20 @@ blogRouter.get('/:id', async (req, res) => {
 })
 
 blogRouter.delete('/:id', async (req, res, next) => {
-  // const blog = await Blog.findById(req.params.id)
-  // const decodedToken = jwt.verify(req.token, process.env.SECRET)
-  // if (!req.token || !decodedToken.id) {
-  //   return res.status(401).json({ error:'token missing or invalid' })
-  // }
-  // if (blog.user.toString() !== decodedToken.id.toString()) {
-  //   await Blog.findByIdAndRemove(req.params.id)
-  //   res.status(204).end()
-  // } else {
-  //   return res.status(401).json({ error: 'a blog can only be deleted by the user who added the blog' })
-  // }
   const blog = await Blog.findById(req.params.id)
-  if(blog){
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if(blog && blog.user.toString() === decodedToken.id.toString()){
     await Blog.findByIdAndDelete(req.params.id)
     res.status(204).end()
   } else {
-    res.status(404)
+    res.status(401).send({ error: 'Not authorized' })
+
   }
 })
 
 blogRouter.put('/:id', async(req, res) => {
   const entry = req.body
-  const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, entry, { new: true })
+  const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, entry, { new: true }).populate('user', { username: 1, name: 1 })
   res.status(200).json(updatedBlog)
 })
